+++ test-templates/src/curves.rs
@@ -1,8 +1,9 @@
 #![allow(unused)]
 use ark_ec::{
-    short_weierstrass_jacobian::GroupAffine, twisted_edwards_extended::GroupProjective,
-    wnaf::WnafContext, AffineCurve, MontgomeryModelParameters, ProjectiveCurve, SWModelParameters,
-    TEModelParameters,
+    short_weierstrass::{Affine, SWCurveConfig},
+    twisted_edwards::{MontCurveConfig, Projective, TECurveConfig},
+    wnaf::WnafContext,
+    AffineCurve, ProjectiveCurve,
 };
 use ark_ff::{Field, One, PrimeField, UniformRand, Zero};
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize, SWFlags, SerializationError};
@@ -292,6 +293,83 @@ pub fn curve_tests<G: ProjectiveCurve>() {
         assert_eq!(b, c);
     }
 
+    {
+        let mut rng = ark_std::test_rng();
+        let a = G::rand(&mut rng);
+        let mut b = G::rand(&mut rng);
+        let zero = G::zero();
+        let fr_zero = G::ScalarField::zero();
+        let fr_one = G::ScalarField::one();
+        let fr_two = fr_one + &fr_one;
+
+        assert_eq!(zero, zero);
+        assert_eq!(zero.is_zero(), true);
+        assert_eq!(a.mul(&fr_one.into_bigint()), a);
+        assert_eq!(a.mul(&fr_two.into_bigint()), a + &a);
+        assert_eq!(a.mul(&fr_zero.into_bigint()), zero);
+        assert_eq!(a.mul(&fr_zero.into_bigint()) - &a, -a);
+        assert_eq!(a.mul(&fr_one.into_bigint()) - &a, zero);
+        assert_eq!(a.mul(&fr_two.into_bigint()) - &a, a);
+
+        // a == a
+        assert_eq!(a, a);
+        // a + 0 = a
+        assert_eq!(a + &zero, a);
+        // a - 0 = a
+        assert_eq!(a - &zero, a);
+        // a - a = 0
+        assert_eq!(a - &a, zero);
+        // 0 - a = -a
+        assert_eq!(zero - &a, -a);
+        // a.double() = a + a
+        assert_eq!(a.double(), a + &a);
+        // b.double() = b + b
+        assert_eq!(b.double(), b + &b);
+        // a + b = b + a
+        assert_eq!(a + &b, b + &a);
+        // a - b = -(b - a)
+        assert_eq!(a - &b, -(b - &a));
+        // (a + b) + a = a + (b + a)
+        assert_eq!((a + &b) + &a, a + &(b + &a));
+        // (a + b).double() = (a + b) + (b + a)
+        assert_eq!((a + &b).double(), (a + &b) + &(b + &a));
+
+        // Check that double_in_place and double give the same result
+        let original_b = b;
+        b.double_in_place();
+        assert_eq!(original_b.double(), b);
+
+        let fr_rand1 = G::ScalarField::rand(&mut rng);
+        let fr_rand2 = G::ScalarField::rand(&mut rng);
+        let a_rand1 = a.mul(&fr_rand1.into_bigint());
+        let a_rand2 = a.mul(&fr_rand2.into_bigint());
+        let fr_three = fr_two + &fr_rand1;
+        let a_two = a.mul(&fr_two.into_bigint());
+        assert_eq!(a_two, a.double(), "(a * 2)  != a.double()");
+        let a_six = a.mul(&(fr_three * &fr_two).into_bigint());
+        assert_eq!(
+            a_two.mul(&fr_three.into_bigint()),
+            a_six,
+            "(a * 2) * 3 != a * (2 * 3)"
+        );
+
+        assert_eq!(
+            a_rand1.mul(&fr_rand2.into_bigint()),
+            a_rand2.mul(&fr_rand1.into_bigint()),
+            "(a * r1) * r2 != (a * r2) * r1"
+        );
+        assert_eq!(
+            a_rand2.mul(&fr_rand1.into_bigint()),
+            a.mul(&(fr_rand1 * &fr_rand2).into_bigint()),
+            "(a * r2) * r1 != a * (r1 * r2)"
+        );
+        assert_eq!(
+            a_rand1.mul(&fr_rand2.into_bigint()),
+            a.mul(&(fr_rand1 * &fr_rand2).into_bigint()),
+            "(a * r1) * r2 != a * (r1 * r2)"
+        );
+    }
+
     random_addition_test::<G>();
     random_multiplication_test::<G>();
     random_doubling_test::<G>();
@@ -299,22 +377,22 @@ pub fn curve_tests<G: ProjectiveCurve>() {
     random_transformation_test::<G>();
 }
 
-pub fn sw_tests<P: SWModelParameters>() {
+pub fn sw_tests<P: SWCurveConfig>() {
     sw_curve_serialization_test::<P>();
     sw_from_random_bytes::<P>();
     sw_affine_sum_test::<P>();
     sw_cofactor_clearing_test::<P>();
 }
 
-pub fn sw_from_random_bytes<P: SWModelParameters>() {
-    use ark_ec::models::short_weierstrass_jacobian::{GroupAffine, GroupProjective};
+pub fn sw_from_random_bytes<P: SWCurveConfig>() {
+    use ark_ec::models::short_weierstrass::{Affine, Projective};
 
-    let buf_size = GroupAffine::<P>::zero().serialized_size();
+    let buf_size = Affine::<P>::zero().serialized_size();
 
     let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
-        let a = GroupProjective::<P>::rand(&mut rng);
+        let a = Projective::<P>::rand(&mut rng);
         let mut a = a.into_affine();
         {
             let mut serialized = vec![0; buf_size];
@@ -322,22 +400,22 @@ pub fn sw_from_random_bytes<P: SWModelParameters>() {
             a.serialize(&mut cursor).unwrap();
 
             let mut cursor = Cursor::new(&serialized[..]);
-            let p1 = GroupAffine::<P>::deserialize(&mut cursor).unwrap();
-            let p2 = GroupAffine::<P>::from_random_bytes(&serialized).unwrap();
+            let p1 = Affine::<P>::deserialize(&mut cursor).unwrap();
+            let p2 = Affine::<P>::from_random_bytes(&serialized).unwrap();
             assert_eq!(p1, p2);
         }
     }
 }
 
-pub fn sw_curve_serialization_test<P: SWModelParameters>() {
-    use ark_ec::models::short_weierstrass_jacobian::{GroupAffine, GroupProjective};
+pub fn sw_curve_serialization_test<P: SWCurveConfig>() {
+    use ark_ec::models::short_weierstrass::{Affine, Projective};
 
-    let buf_size = GroupAffine::<P>::zero().serialized_size();
+    let buf_size = Affine::<P>::zero().serialized_size();
 
     let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
-        let a = GroupProjective::<P>::rand(&mut rng);
+        let a = Projective::<P>::rand(&mut rng);
         let mut a = a.into_affine();
         {
             let mut serialized = vec![0; buf_size];
@@ -345,7 +423,7 @@ pub fn sw_curve_serialization_test<P: SWModelParameters>() {
             a.serialize(&mut cursor).unwrap();
 
             let mut cursor = Cursor::new(&serialized[..]);
-            let b = GroupAffine::<P>::deserialize(&mut cursor).unwrap();
+            let b = Affine::<P>::deserialize(&mut cursor).unwrap();
             assert_eq!(a, b);
         }
 
@@ -355,22 +433,22 @@ pub fn sw_curve_serialization_test<P: SWModelParameters>() {
             let mut cursor = Cursor::new(&mut serialized[..]);
             a.serialize(&mut cursor).unwrap();
             let mut cursor = Cursor::new(&serialized[..]);
-            let b = GroupAffine::<P>::deserialize(&mut cursor).unwrap();
+            let b = Affine::<P>::deserialize(&mut cursor).unwrap();
             assert_eq!(a, b);
         }
 
         {
-            let a = GroupAffine::<P>::zero();
+            let a = Affine::<P>::zero();
             let mut serialized = vec![0; buf_size];
             let mut cursor = Cursor::new(&mut serialized[..]);
             a.serialize(&mut cursor).unwrap();
             let mut cursor = Cursor::new(&serialized[..]);
-            let b = GroupAffine::<P>::deserialize(&mut cursor).unwrap();
+            let b = Affine::<P>::deserialize(&mut cursor).unwrap();
             assert_eq!(a, b);
         }
 
         {
-            let a = GroupAffine::<P>::zero();
+            let a = Affine::<P>::zero();
             let mut serialized = vec![0; buf_size - 1];
             let mut cursor = Cursor::new(&mut serialized[..]);
             a.serialize(&mut cursor).unwrap_err();
@@ -379,7 +457,7 @@ pub fn sw_curve_serialization_test<P: SWModelParameters>() {
         {
             let serialized = vec![0; buf_size - 1];
             let mut cursor = Cursor::new(&serialized[..]);
-            GroupAffine::<P>::deserialize(&mut cursor).unwrap_err();
+            Affine::<P>::deserialize(&mut cursor).unwrap_err();
         }
 
         {
@@ -388,7 +466,7 @@ pub fn sw_curve_serialization_test<P: SWModelParameters>() {
             a.serialize_uncompressed(&mut cursor).unwrap();
 
             let mut cursor = Cursor::new(&serialized[..]);
-            let b = GroupAffine::<P>::deserialize_uncompressed(&mut cursor).unwrap();
+            let b = Affine::<P>::deserialize_uncompressed(&mut cursor).unwrap();
             assert_eq!(a, b);
         }
 
@@ -398,35 +476,35 @@ pub fn sw_curve_serialization_test<P: SWModelParameters>() {
             let mut cursor = Cursor::new(&mut serialized[..]);
             a.serialize_uncompressed(&mut cursor).unwrap();
             let mut cursor = Cursor::new(&serialized[..]);
-            let b = GroupAffine::<P>::deserialize_uncompressed(&mut cursor).unwrap();
+            let b = Affine::<P>::deserialize_uncompressed(&mut cursor).unwrap();
             assert_eq!(a, b);
         }
 
         {
-            let a = GroupAffine::<P>::zero();
+            let a = Affine::<P>::zero();
             let mut serialized = vec![0; a.uncompressed_size()];
             let mut cursor = Cursor::new(&mut serialized[..]);
             a.serialize_uncompressed(&mut cursor).unwrap();
             let mut cursor = Cursor::new(&serialized[..]);
-            let b = GroupAffine::<P>::deserialize_uncompressed(&mut cursor).unwrap();
+            let b = Affine::<P>::deserialize_uncompressed(&mut cursor).unwrap();
             assert_eq!(a, b);
         }
     }
 }
 
-pub fn sw_affine_sum_test<P: SWModelParameters>() {
-    use ark_ec::models::short_weierstrass_jacobian::{GroupAffine, GroupProjective};
+pub fn sw_affine_sum_test<P: SWCurveConfig>() {
+    use ark_ec::models::short_weierstrass::{Affine, Projective};
 
     let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
         let mut test_vec = Vec::new();
         for _ in 0..10 {
-            test_vec.push(GroupProjective::<P>::rand(&mut rng).into_affine());
+            test_vec.push(Projective::<P>::rand(&mut rng).into_affine());
         }
 
-        let sum_computed: GroupAffine<P> = test_vec.iter().sum();
-        let mut sum_expected = GroupAffine::zero();
+        let sum_computed: Affine<P> = test_vec.iter().sum();
+        let mut sum_expected = Affine::zero();
         for p in test_vec.iter() {
             sum_expected += p;
         }
@@ -435,11 +513,11 @@ pub fn sw_affine_sum_test<P: SWModelParameters>() {
     }
 }
 
-fn sw_cofactor_clearing_test<P: SWModelParameters>() {
+fn sw_cofactor_clearing_test<P: SWCurveConfig>() {
     let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
-        let a = GroupAffine::<P>::rand(&mut rng);
+        let a = Affine::<P>::rand(&mut rng);
         let b = a.clear_cofactor();
         assert!(b.is_in_correct_subgroup_assuming_on_curve());
     }
@@ -447,7 +525,7 @@ fn sw_cofactor_clearing_test<P: SWModelParameters>() {
 
 pub fn montgomery_conversion_test<P>()
 where
-    P: TEModelParameters,
+    P: TECurveConfig,
 {
     // A = 2 * (a + d) / (a - d)
     let a = P::BaseField::one().double()
@@ -456,11 +534,11 @@ where
     // B = 4 / (a - d)
     let b = P::BaseField::one().double().double() * &(P::COEFF_A - &P::COEFF_D).inverse().unwrap();
 
-    assert_eq!(a, P::MontgomeryModelParameters::COEFF_A);
-    assert_eq!(b, P::MontgomeryModelParameters::COEFF_B);
+    assert_eq!(a, P::MontCurveConfig::COEFF_A);
+    assert_eq!(b, P::MontCurveConfig::COEFF_B);
 }
 
-pub fn edwards_tests<P: TEModelParameters>()
+pub fn edwards_tests<P: TECurveConfig>()
 where
     P::BaseField: PrimeField,
 {
@@ -469,18 +547,18 @@ where
     edwards_cofactor_clearing_test::<P>();
 }
 
-pub fn edwards_from_random_bytes<P: TEModelParameters>()
+pub fn edwards_from_random_bytes<P: TECurveConfig>()
 where
     P::BaseField: PrimeField,
 {
-    use ark_ec::models::twisted_edwards_extended::{GroupAffine, GroupProjective};
+    use ark_ec::models::twisted_edwards::{Affine, Projective};
 
-    let buf_size = GroupAffine::<P>::zero().serialized_size();
+    let buf_size = Affine::<P>::zero().serialized_size();
 
     let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
-        let a = GroupProjective::<P>::rand(&mut rng);
+        let a = Projective::<P>::rand(&mut rng);
         let mut a = a.into_affine();
         {
             let mut serialized = vec![0; buf_size];
@@ -488,15 +566,15 @@ where
             a.serialize(&mut cursor).unwrap();
 
             let mut cursor = Cursor::new(&serialized[..]);
-            let p1 = GroupAffine::<P>::deserialize(&mut cursor).unwrap();
-            let p2 = GroupAffine::<P>::from_random_bytes(&serialized).unwrap();
+            let p1 = Affine::<P>::deserialize(&mut cursor).unwrap();
+            let p2 = Affine::<P>::from_random_bytes(&serialized).unwrap();
             assert_eq!(p1, p2);
         }
     }
 
     for _ in 0..ITERATIONS {
         let mut biginteger =
-            <<GroupAffine<P> as AffineCurve>::BaseField as PrimeField>::BigInt::rand(&mut rng);
+            <<Affine<P> as AffineCurve>::BaseField as PrimeField>::BigInt::rand(&mut rng);
         let mut bytes = {
             let mut result = vec![0u8; biginteger.serialized_size()];
             biginteger
@@ -504,24 +582,24 @@ where
                 .unwrap();
             result
         };
-        let mut g = GroupAffine::<P>::from_random_bytes(&bytes);
+        let mut g = Affine::<P>::from_random_bytes(&bytes);
         while g.is_none() {
             bytes.iter_mut().for_each(|i| *i = i.wrapping_sub(1));
-            g = GroupAffine::<P>::from_random_bytes(&bytes);
+            g = Affine::<P>::from_random_bytes(&bytes);
         }
         let _g = g.unwrap();
     }
 }
 
-pub fn edwards_curve_serialization_test<P: TEModelParameters>() {
-    use ark_ec::models::twisted_edwards_extended::{GroupAffine, GroupProjective};
+pub fn edwards_curve_serialization_test<P: TECurveConfig>() {
+    use ark_ec::models::twisted_edwards::{Affine, Projective};
 
-    let buf_size = GroupAffine::<P>::zero().serialized_size();
+    let buf_size = Affine::<P>::zero().serialized_size();
 
     let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
-        let a = GroupProjective::<P>::rand(&mut rng);
+        let a = Projective::<P>::rand(&mut rng);
         let a = a.into_affine();
         {
             let mut serialized = vec![0; buf_size];
@@ -529,22 +607,22 @@ pub fn edwards_curve_serialization_test<P: TEModelParameters>() {
             a.serialize(&mut cursor).unwrap();
 
             let mut cursor = Cursor::new(&serialized[..]);
-            let b = GroupAffine::<P>::deserialize(&mut cursor).unwrap();
+            let b = Affine::<P>::deserialize(&mut cursor).unwrap();
             assert_eq!(a, b);
         }
 
         {
-            let a = GroupAffine::<P>::zero();
+            let a = Affine::<P>::zero();
             let mut serialized = vec![0; buf_size];
             let mut cursor = Cursor::new(&mut serialized[..]);
             a.serialize(&mut cursor).unwrap();
             let mut cursor = Cursor::new(&serialized[..]);
-            let b = GroupAffine::<P>::deserialize(&mut cursor).unwrap();
+            let b = Affine::<P>::deserialize(&mut cursor).unwrap();
             assert_eq!(a, b);
         }
 
         {
-            let a = GroupAffine::<P>::zero();
+            let a = Affine::<P>::zero();
             let mut serialized = vec![0; buf_size - 1];
             let mut cursor = Cursor::new(&mut serialized[..]);
             a.serialize(&mut cursor).unwrap_err();
@@ -553,7 +631,7 @@ pub fn edwards_curve_serialization_test<P: TEModelParameters>() {
         {
             let serialized = vec![0; buf_size - 1];
             let mut cursor = Cursor::new(&serialized[..]);
-            GroupAffine::<P>::deserialize(&mut cursor).unwrap_err();
+            Affine::<P>::deserialize(&mut cursor).unwrap_err();
         }
 
         {
@@ -562,27 +640,27 @@ pub fn edwards_curve_serialization_test<P: TEModelParameters>() {
             a.serialize_uncompressed(&mut cursor).unwrap();
 
             let mut cursor = Cursor::new(&serialized[..]);
-            let b = GroupAffine::<P>::deserialize_uncompressed(&mut cursor).unwrap();
+            let b = Affine::<P>::deserialize_uncompressed(&mut cursor).unwrap();
             assert_eq!(a, b);
         }
 
         {
-            let a = GroupAffine::<P>::zero();
+            let a = Affine::<P>::zero();
             let mut serialized = vec![0; a.uncompressed_size()];
             let mut cursor = Cursor::new(&mut serialized[..]);
             a.serialize_uncompressed(&mut cursor).unwrap();
             let mut cursor = Cursor::new(&serialized[..]);
-            let b = GroupAffine::<P>::deserialize_uncompressed(&mut cursor).unwrap();
+            let b = Affine::<P>::deserialize_uncompressed(&mut cursor).unwrap();
             assert_eq!(a, b);
         }
     }
 }
 
-fn edwards_cofactor_clearing_test<P: TEModelParameters>() {
+fn edwards_cofactor_clearing_test<P: TECurveConfig>() {
     let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
-        let a = GroupProjective::<P>::rand(&mut rng).into_affine();
+        let a = Projective::<P>::rand(&mut rng).into_affine();
         let b = a.clear_cofactor();
         assert!(b.is_in_correct_subgroup_assuming_on_curve());
     }